fixed the issue for the unit decimal values, change the double to int datatype for quantity in sell lines and also change the round up value to 1 dec point  9:13pm, 03-Oct-2023 by afaq

+++ lib/Controllers/ProductController/product_cart_controller.dart
@@ -49,7 +49,7 @@ class ProductCartController extends GetxController {
     shippingChargeCtrl.clear();
   }
 
-  double productTotalAmount(String? price, int quantity) {
+  double productTotalAmount(String? price, double quantity) {
     return double.parse('${price ?? 0.0}') * quantity;
   }
